+++ device/CMakeLists.txt
@@ -113,7 +113,9 @@ if(TT_UMD_BUILD_SIMULATION)
     target_sources(
         device
         PRIVATE
-            simulation/simulation_device.cpp
+            simulation/simulation_chip.cpp
+            simulation/tt_simulation_chip.cpp
+            simulation/rtl_simulation_chip.cpp
             simulation/simulation_host.cpp
             ${FBS_GENERATED_HEADER}
     )
@@ -164,7 +166,9 @@ if(CMAKE_VERSION VERSION_GREATER_EQUAL 3.23)
                 api/umd/device/tt_io.hpp
                 api/umd/device/types/tensix_soft_reset_options.hpp
                 api/umd/device/types/risc_type.hpp
-                api/umd/device/simulation/simulation_device.hpp
+                api/umd/device/simulation/simulation_chip.hpp
+                api/umd/device/simulation/tt_simulation_chip.hpp
+                api/umd/device/simulation/rtl_simulation_chip.hpp
                 api/umd/device/simulation/simulation_host.hpp
                 api/umd/device/soc_descriptor.hpp
                 api/umd/device/tt_xy_pair.h
@@ -182,7 +186,6 @@ if(CMAKE_VERSION VERSION_GREATER_EQUAL 3.23)
                 # To be removed
                 api/umd/device/hugepage.h
                 api/umd/device/tt_core_coordinates.h
-                api/umd/device/tt_simulation_device.h
                 api/umd/device/tt_cluster_descriptor.h
                 api/umd/device/tt_soc_descriptor.h
                 api/umd/device/wormhole_implementation.h
