--- device/api/umd/device/tt_simulation_device.h
@@ -1,7 +1,0 @@
-/*
- * SPDX-FileCopyrightText: (c) 2025 Tenstorrent Inc.
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-// TODO: To be removed once clients switch to including the other header.
-#include "umd/device/simulation/simulation_device.hpp"
